+++ Engine/PanzerBot.h
@@ -5,7 +5,7 @@ class PanzerBot : public Panzer {
 public :
 	PanzerBot() {};
 	void logic() ;
-	PanzerBot(int x, int y, int dir, int vel, void*_p, void* _map,int type);
+	PanzerBot(float x, float y, int dir, float vel, void*_p, void* _map,int type);
 	void del() { IsAlive = 0; }
 
 };
